requirements.txt
asgiref==3.7.2
cffi==1.15.1
Django==3.2.21
django-appconf==1.0.5
django-filter==23.3
djangorestframework==3.14.0
generics==6.0.0
pyasn1==0.5.0
pycparser==2.21
pycrypto==2.6.1
pytz==2023.3.post1
sqlparse==0.4.4
typing-extensions==4.7.1
